move error handle of response to response class
